Move rtt into own module
